fix(actions): consolidate helpers onto harness types, remove op-e2e refs
Free-floating helper functions in action test files are moved onto framework
types: create_l2_source and l1_info_from_block onto ActionTestHarness,
enqueue_* log-building methods onto L1Miner, and mainnet() onto
TestRollupConfigBuilder. All op-e2e cross-references are removed from
comments throughout actions/ as Base's test framework is independent.
